fix: [BE/TodoRepo] Todo 달성도 0으로 표시되는 문제 해결

+++ Backend/src/main/java/com/ssafy/backend/mypage/domain/dto/CafeLiveRespDto.java
@@ -17,7 +17,7 @@ public class CafeLiveRespDto {
 
     private int visitedAt; // 입실 일자 (yyyy-MM-dd)
     private int accTime; // 누적 시간
-    private int todoAchievementRate; // 달성도
+    private Double todoAchievementRate; // 달성도
     private CafeNameAndBrandDto cafeInfo; // name, brand
     private String fortuneMsg;
